requirements.txt
djangorestframework
django-rest-knox
django-extensions
Django==2.0.12
celery==4.3.0
Unipath==1.1
django-celery-beat==1.5.0
psycopg2==2.7.7
requests==2.22.0